requirements.txt
# =============================
# MCP Framework - Core Dependencies
# =============================

# Web API
fastapi>=0.121.0
uvicorn>=0.38.0

# Core logic and graph orchestration
langchain-core>=1.0.0
langgraph>=1.0.0

# Storage / Memory
redis>=7.0.0

# Config, typing & serialization
pydantic>=2.12.0
python-dotenv>=1.2.0
PyYAML>=6.0.0
httpx>=0.28.0
requests>=2.32.0
tenacity>=9.1.0
orjson>=3.11.0
